requirements.txt
python-telegram-bot[job-queue]==21.6
httpx==0.27.2
aiosqlite==0.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
